happy case

+++ service/src/main/java/cz/muni/fi/pa165/mushrooms/service/facade/MushroomFacadeImpl.java
@@ -48,18 +48,30 @@ public class MushroomFacadeImpl implements MushroomFacade {
 
     @Override
     public List<MushroomDTO> findByMushroomType(MushroomType mushroomType) {
+        if (mushroomType == null) {
+            throw new IllegalArgumentException("No mushroom type at finding mushroom by type.");
+        }
         List<Mushroom> mushrooms = service.findByMushroomType(mushroomType);
         return beanMappingService.mapTo(mushrooms, MushroomDTO.class);
     }
 
     @Override
     public List<MushroomDTO> findByIntervalOfOccurrence(String fromMonth, String toMonth) {   ////// TODO String - Date
+        if (fromMonth == null) {
+            throw new IllegalArgumentException("Parameter fromMonth is null");
+        }
+        if (toMonth == null) {
+            throw new IllegalArgumentException("Parameter toMonth is null");
+        }
         List<Mushroom> mushrooms = service.findByIntervalOfOccurrence(fromMonth, toMonth);
         return beanMappingService.mapTo(mushrooms, MushroomDTO.class);
     }
 
     @Override
     public MushroomDTO createMushroom(MushroomDTO mushroom) {
+        if (mushroom == null) {
+            throw new IllegalArgumentException("No mushroom at mushroom create.");
+        }
         Mushroom mappedMushroom = beanMappingService.mapTo(mushroom, Mushroom.class);
         //set mushroom
         mappedMushroom.setName(mushroom.getName());
@@ -73,6 +85,9 @@ public class MushroomFacadeImpl implements MushroomFacade {
 
     @Override   //InvocationTargetException
     public boolean deleteMushroom(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Null id at mushroom delete.");
+        }
         Mushroom mushroom = beanMappingService.mapTo(findMushroomById(id), Mushroom.class);
         service.deleteMushroom(mushroom);
         return true;
@@ -80,6 +95,9 @@ public class MushroomFacadeImpl implements MushroomFacade {
 
     @Override
     public MushroomDTO updateMushroom(MushroomDTO mushroom) {
+        if (mushroom == null) {
+            throw new IllegalArgumentException("No mushroom at mushroom update.");
+        }
         Mushroom entity = service.findMushroomById(mushroom.getId());
         entity.setName(mushroom.getName());
         entity.setType(mushroom.getType());
